moved invisible image input into QRCodeScanner
